Reorganisation du code
