+++ src/pf/yozzman/risk/util/BarreChargement.java
@@ -1,4 +1,4 @@
-package pf.yozzman.risk;
+package pf.yozzman.risk.util;
 
 public class BarreChargement {
